remove etcd2 client from server
